+++ jawa/boba-fett/src/main/java/ru/prohor/universe/bobafett/feature/currency/CurrencyLayerApiService.java
@@ -65,5 +65,5 @@ public class CurrencyLayerApiService {
                 .toList();
     }
 
-    private record CurrencyResponse(Map<String, Float> quotes) {}
+    private record CurrencyResponse(Map<String, Double> quotes) {}
 }
